Adds float capability to integers

+++ db/migrate/20140205174054_create_schools.rb
@@ -2,9 +2,9 @@ class CreateSchools < ActiveRecord::Migration
   def change
     create_table :schools do |t|
       t.string :title
-      t.integer :left_operand
+      t.decimal :left_operand
       t.string :operator
-      t.integer :right_operand
+      t.decimal :right_operand
 
       t.timestamps
     end
